[Trainer] Add failing test with cross_entropy loss function
Added a test which reproduces issue during running training with
cross_enropy loss function.
This test was failed with the following error:
```
Check failed: (idx < data_.size() && data_[idx].second != 0) is false: Attribute FPrimalGradient has not been registered for relax.nn.nll_loss
```

To quick fix for this issue (just to check that cross_entropy_loss
function works), added definition of gradient function for
`nll_loss`. But for now it reports the following error:
```
TVMError: Data types int64 and float32 must be equal for binary operators
```

I suppose it is happened because NLLLoss function takes float array for
predictions and integer for targets. If I try to use float datatype for
targets then I get the following error:
```
NLLLoss expects the dtype of targets to be int/uint. However, the dtype of targets is float32
```

The issue with data types for gradient function should be debugged and
resolved. I hope after that it will be possible to use
`CrossEntropyLoss` function for training.